+++ src/Paginator/KeysetPaginator.php
@@ -13,6 +13,7 @@ use Yiisoft\Data\Reader\Filter\GreaterThanOrEqual;
 use Yiisoft\Data\Reader\Filter\LessThan;
 use Yiisoft\Data\Reader\Filter\LessThanOrEqual;
 use Yiisoft\Data\Reader\FilterableDataInterface;
+use Yiisoft\Data\Reader\LimitableDataInterface;
 use Yiisoft\Data\Reader\ReadableDataInterface;
 use Yiisoft\Data\Reader\Sort;
 use Yiisoft\Data\Reader\SortableDataInterface;
@@ -49,7 +50,7 @@ final class KeysetPaginator implements PaginatorInterface
     /**
      * Data reader being paginated.
      *
-     * @psalm-var ReadableDataInterface<TKey, TValue>&FilterableDataInterface&SortableDataInterface
+     * @psalm-var ReadableDataInterface<TKey, TValue>&LimitableDataInterface&FilterableDataInterface&SortableDataInterface
      */
     private ReadableDataInterface $dataReader;
 
@@ -82,7 +83,7 @@ final class KeysetPaginator implements PaginatorInterface
 
     /**
      * @param ReadableDataInterface $dataReader Data reader being paginated.
-     * @psalm-param ReadableDataInterface<TKey, TValue>&FilterableDataInterface&SortableDataInterface $dataReader
+     * @psalm-param ReadableDataInterface<TKey, TValue>&LimitableDataInterface&FilterableDataInterface&SortableDataInterface $dataReader
      * @psalm-suppress DocblockTypeContradiction Needed to allow validating `$dataReader`
      */
     public function __construct(ReadableDataInterface $dataReader)
@@ -98,6 +99,13 @@ final class KeysetPaginator implements PaginatorInterface
             throw new InvalidArgumentException(sprintf(
                 'Data reader should implement "%s" to be used with keyset paginator.',
                 SortableDataInterface::class,
+            ));
+        }
+
+        if (!$dataReader instanceof LimitableDataInterface) {
+            throw new InvalidArgumentException(sprintf(
+                'Data reader should implement "%s" to be used with keyset paginator.',
+                LimitableDataInterface::class,
             ));
         }
 
@@ -183,6 +191,15 @@ final class KeysetPaginator implements PaginatorInterface
         }
 
         return $this->readCache = $data;
+    }
+
+    public function readOne(): array|object|null
+    {
+        foreach ($this->read() as $item) {
+            return $item;
+        }
+
+        return null;
     }
 
     public function getPageSize(): int
@@ -289,7 +306,7 @@ final class KeysetPaginator implements PaginatorInterface
     }
 
     /**
-     * @psalm-param ReadableDataInterface<TKey, TValue>&FilterableDataInterface&SortableDataInterface $dataReader
+     * @psalm-param ReadableDataInterface<TKey, TValue>&LimitableDataInterface&FilterableDataInterface&SortableDataInterface $dataReader
      */
     private function previousPageExist(ReadableDataInterface $dataReader, Sort $sort): bool
     {